add more filters to query
